requirements.txt
axios==0.4.0
celery==5.3.6
cors==1.0.1
Flask==3.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
redis==5.0.1
SQLAlchemy==2.0.25

